backend/requirements.txt
# AgentGround — Python Dependencies
# Python 3.10+  |  Install: pip install -r backend/requirements.txt

# UI
streamlit==1.35.0

# Agent Framework
langchain==0.2.16
langchain-community==0.2.16
langchain-core<0.3.0
# Note: xAI, DeepSeek, and Ollama are supported natively via langchain-openai
# because they provide OpenAI-compatible endpoints.
langchain-openai==0.1.23

# Database
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.3

# Environment variables
python-dotenv==1.0.1

# Web Search (requires TAVILY_API_KEY in .env)
tavily-python==0.3.3

# Google Gemini support
langchain-google-genai==1.0.10
google-generativeai==0.7.2
